+++ src/config/pipe.config.ts
@@ -1,4 +1,4 @@
-import { validationPipe } from '../pipe/validation.pipe';
+import { validationPipe } from '../common/pipe/validation.pipe';
 
 export const pipeConfig = (app) => {
   validationPipe(app);
